Potential fix for pull request finding

+++ pixeltable/catalog/catalog.py
@@ -1154,7 +1154,7 @@ class Catalog:
             update_status['cascade_row_count_stats']['num_excs'].astext.cast(sql.Integer), 0
         )
         errors = sql.func.coalesce(
-            sql.cast(sql.func.sum(row_count_excs + cascade_row_count_excs), sql.Integer), 0
+            sql.cast(sql.func.sum(row_count_excs + cascade_row_count_excs), sql.BigInteger), 0
         ).label('errors')
         stmt = sql.select(schema.TableVersion.tbl_id, errors).group_by(schema.TableVersion.tbl_id)
         rows = get_runtime().conn.execute(stmt).all()
